整数溢出修改
# message auto-generated for no-merge-commit merge:
merge ******_int_flow into master

整数溢出安全排查

Created-by: insightDeveloper

Match-id-f21895490eb15c46810fb1ee8252c4a9ef22ef3a

+++ server/src/modules/communication/parser/ClusterDef.h
@@ -81,7 +81,7 @@ struct ClusterBaseInfo {
     std::string steps;
     std::string ppStages;
     std::string stages;
-    long collectStartTime;
+    long long collectStartTime;
     double collectDuration;
     long long dataSize  =  0;
     ParallelStrategyConfig config;
